ws/python2/requirements.txt
websocket-client==0.57.0